refactor(building): split tests.rs into tranche-scoped modules; tighten emit passes
- Split crates/mogen-dsl/src/lower/building/tests.rs (1204 lines) into
  per-tranche modules under tests/ (single_storey, multi_storey, layouts,
  roof_and_cellar) plus a shared mod.rs with helpers and SRC fixtures.
  All 73 building tests still pass.
- Reshape emit/circulation.rs to align stair landings and elevator
  approaches with the corridor strip; fix interior wall rotation for
  perpendicular runs.
- Tighten emit/openings.rs and emit/shell.rs for the new circulation
  geometry; refresh stdlib/stair_simple.mog to match.
